Create FlatfileDatabase class.
